Refactor the components
