+++ app/goal/_components/ActionModal/TabContentSection.tsx
@@ -1,0 +1,55 @@
+import React from "react";
+import { Text, StyleSheet } from "react-native";
+import { WhoTabContent } from "./WhoTabContent";
+import { TabType, Person } from "./types";
+
+interface TabContentSectionProps {
+  activeTab: TabType;
+  people: Person[];
+  selectedPersonId: string | null;
+  onPersonSelect: (personId: string) => void;
+}
+
+export function TabContentSection({
+  activeTab,
+  people,
+  selectedPersonId,
+  onPersonSelect,
+}: TabContentSectionProps) {
+  switch (activeTab) {
+    case "WHY":
+      return (
+        <Text style={styles.contentDescription}>
+          Your motivation for this goal
+        </Text>
+      );
+
+    case "WHEN":
+      return (
+        <Text style={styles.contentDescription}>
+          Select your goal date below
+        </Text>
+      );
+
+    case "WHO":
+      return (
+        <WhoTabContent
+          people={people}
+          selectedPersonId={selectedPersonId}
+          onPersonSelect={onPersonSelect}
+        />
+      );
+
+    default:
+      return null;
+  }
+}
+
+const styles = StyleSheet.create({
+  contentDescription: {
+    marginTop: 8,
+    fontSize: 14,
+    color: "#666",
+    textAlign: "center",
+  },
+});
