+++ app/goal/_components/ActionModal/index.ts
@@ -1,0 +1,3 @@
+export { ActionModal } from "./ActionModal";
+export type { TabType, Person, ActionModalProps } from "./types";
+export { TABS, MOCK_PEOPLE } from "./constants";
